Set executable bit on common scripting language files
If you encrypt (or decrypt) a file with a common scripting extension
(eg. `.js`, `.sh`) then `git-cipher` will set its executable bit
automatically. We didn't used to do this, but that meant that if you
went back and manually set the permissions then you would bump the
modification time of the file, causing it to be needlessly reencrypted
on the next run of `git-cipher encrypt`.

Other approaches that were considered and discarded were:

1. Make the behavior optional, via a switch or configuration option
   (discarded due to needless complexity).
2. Encode the need to set the bit in the encrypted file's own
   permissions (discarded because that means setting the bit on a file
   that is not executable, which is misleading).
3. Encode the need to set the bit in the encrypted file's internal
   structure: effectively, instead of storing the GPG output, we'd store
   a YAML file containing metadata and the output (discarded because of
   compatibility concerns and complexity; would make many subcommands in
   the app more complicated).
4. Encode the need to set the bit in the encrypted file's name
   (discarded because it's ugly and would still add a fair bit of
   complexity).

So in the end have taken what I think is the pragmatic approach.