feat: support encryption with multiple keys
The motivation, as explained here[0], is to have these keys and subkeys:

1.  Personal (signing) key: alice@home.example.net, not used by
    git-cipher
    -   Subkey for encryption (added by GPG): used for encryption
    -   Any other subkeys (optionally) added by user: not used by
        git-cipher
2.  Work (signing) key: alice@work.example.com, not used by git-cipher
    -   Subkey for encryption (added by GPG): used for encryption
    -   Any other subkeys (optionally) added by user: not used by
        git-cipher

The primary keys can have expiry dates or not (I recommend that they
have them, but that you extend them periodically, for reasons that are
too long to go into here).

The subkeys should have expiry dates. When they expire, rotate to new
keys (ie. re-encrypt the files). To make life easy, the expiry dates
should be synchronized.

Having personal and work keys makes it possible to revoke keys,
rotate them independently if desired, or stop using a key upon leaving a
job.

[0]: https://wincent.com/wiki/GPG_key_rotation_notes