fix(unlock): don't blow away worktree changes when nothing is managed
Discovered while testing locally. If you do this:

1. `git-cipher init`
2. `git-cipher unlock`

If you have any local changes, it will tell you to run with `--force` if
you really want to unlock (because it wants to checkout the plaintext
versions of the files). The gotcha is that if you haven't added any
files yet, then the command that runs ends up being (basically):

    git checkout --force HEAD --

instead of:

    git checkout --force HEAD -- file1 file2...

So, it has the effect of blowing away changes in unmanaged files too.