refactor: simplify(?) wrapping logic
- Get rid of ugly `fits` temporary var.
- Move `length` computation up.

It does make the `if` logic a bit more complicated, but overall the code
is more concise. Hopefully it is still clear enough with the comment.

One thing to note: if a word is longer than `width`, it will wind up
with a greater-than-zero penalty, (because `Math.pow(width - length, 2)`
will be a negative number, squared, so positive), but we'll still get
an optimal solution because such words can _only_ ever appear on their
own on a line, so their penalty score is meaningless. In fact, you could
argue that having them have a non-zero penalty is at least conceptually
"nice", because it shows their contribution to "raggedness" on the right
edge.