feat(unlock): teach `unlock` to alert when `init` needs to be run
I don't know if it is a great design to have `init` and `unlock`
commands both be necessary in order to start working in a repo, although
the initial motivation was good (ie. one command to prepare the repo for
operation, and a pair of symmetrical commands — "lock" and "unlock" —
that solely deal with granting and revoking access to secrets).

The problem is even I may forget to run `init` before `unlock`. One
option is to have `unlock` run `init` automatically if needed, but I
don't really like that as `init` does a wee bit too much (like
generating secrets if none exist) for me to feel comfortable with
running it automatically.

So, instead, I'm adding the code in this commit to at least remind the
user to run `init` if they don't appear to have done so yet.

This fixes the problem that you would otherwise see of, for example:

1. Running `git-cipher unlock` in a repo like this one which already has
   a .gitattributes file.
2. Seeing `git status` bomb out with "fatal: examples/empty-file: clean
   filter 'git-cipher' failed" because I forgot to run `git-cipher init`
   (ie. to set up the filters in the config).