fix: always run `yarn` commands from the repo root
This allows you to add git-cipher as a submodule to another repo and
then do something like:

    vendor/git-cipher/bin/git-cipher init

and have it work.

The script will note that `yarn` hasn't run because there is no
integrity file, and then run it for you automatically. In order for that
to actually work as intended, it's important that it run from the right
directory ("vendor/git-cipher" in this case, not the superproject's repo
root).