fix: don't redirect to `/dev/stderr`
This is breaking in Codespaces for some reason. There _is_ a file at
`/dev/stderr` and I can write to it in an interactive shell, but for
some reason when I write to it in this script, I get:

    /dev/stderr: No such device or address

Whatever. Rewrite using the ugly and hard to read syntax that binds to
file descriptor 2. That seems to work.