fix: make `n which` detection work
Because it doesn't exit with a non-zero exit code if a version is
missing. It prints the path to stdout when a version is found, and
prints a diagnostic message to stderr when a version is not found, and
the exit code is zero either way.

So, we look for empty stdout as a way of detecting missing version. This
seemed less fragile than looking for a specific error string on stderr.