fix: address possible issue in wrapping algorithm
This doesn't change the output in any way, because our use case is
simple, but I believe it is more technically correct. (Motivated by an
issue I found with an implementation of the same algorithm that I did at
work, but which had the ability to break on different characters; in
that more complicated use case, the tweak actually does matter and
produces better wrapping).