feat: don't churn blob contents when bumping version number
Having a version-specific "url" field in the ciphertext causes the blobs
to change needlessly just because the package version number changes. In
practice, that doesn't need to happen except for when the "version"
field changes (ie. a new protocol applies).

So, use a stable URL target instead, and update the linked document to
reflect that it covers the current and all prior versions too.

Re-generated the blobs with:

    git add --renormalize examples